pyproject.toml
[project]
name = "daily-slur-meter"
version = "0.1.0"
description = "Analyse movie subtitles for profanity and generate 9:16 rage chart Shorts"
requires-python = ">=3.11"
dependencies = [
    # FastAPI backend
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic>=2.0.0",
    "python-multipart>=0.0.18",
    # Shared pipeline deps
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "pysrt>=1.1.2",
    "requests>=2.31.0",
    "rarfile>=4.1",
    "pandas>=2.1.0",
    "regex>=2023.10.3",
    "matplotlib>=3.8.0",
    "numpy>=1.26.0",
    "pillow>=10.0.0",
    "edge-tts>=6.1.9",
    "tqdm>=4.66.0",
    # Publishing
    "google-api-python-client>=2.108.0",
    "google-auth-httplib2>=0.1.1",
    "google-auth-oauthlib>=1.1.0",
    "playwright>=1.40.0",
    "scipy>=1.17.1",
]

# Optional test deps
[project.optional-dependencies]
test = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "pytest-mock>=3.12",
    "httpx>=0.27.0",
    "ruff>=0.8.0",
]

[project.scripts]
slur-meter = "main:main"

# ─── Tooling ─────────────────────────────
[tool.ruff]
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "N", "B", "SIM"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks integration tests",
    "unit: marks unit tests",
]
